before stiling nav
