+++ frontend/package.json
@@ -13,6 +13,7 @@
     "eslint": "^8.57.1",
     "formik": "^2.4.6",
     "framer-motion": "^12.0.5",
+    "ionicons": "^7.4.0",
     "lucide-react": "^0.474.0",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.8",
