+++ frontend/public/index.html
@@ -10,6 +10,8 @@
       content="A comprehensive platform for managing and optimizing electric vehicle (EV) battery health, energy usage, predictive maintenance, and charging, enhancing the overall EV driving experience."
     />
     <title>BatterySync</title>
+    <script type="module" src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.esm.js"></script>
+    <script nomodule src="https://unpkg.com/ionicons@7.1.0/dist/ionicons/ionicons.js"></script>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
